pytorch/requirements.txt
hydra-core
hydra_colorlog
hydra-nevergrad-sweeper
python-dotenv
torchio
monai
tensorboardX
plotly
multitasking
pytorch-model-summary
